Use shared_ptr for request body
The `req` member data of the class `Request` is defined as a simple
value.

```
struct Request
{
   boost::beast::http::request<bmcweb::HttpBody> req;
...
}
```

This exhibits a few issues.

- If it is moved to a lambda of user authentication async method call,
 the original `completeRequest()` may observe the invalid content of
it [1].  As the result, the reference of `req` in `completeRequest()` is
not possible, and it may cause the troubles (e.g. audit log later), esp
on PATCH..

```
     crow::connections::systemBus->async_method_call(
        [req{std::move(req)}, asyncResp, &rule,
         callback(std::forward<CallbackFn>(callback))](
```

For example, adding a debug statement to print `req.url()` inside
`completeRequest()` may cause the coredump or give the garbled content..

```
    void completeRequest(crow::Response& thisRes)
    {
        BMCWEB_LOG_INFO("TEST: completeRequest req.url={}", req.url());
        ...
    }
```

- It may be copied (e.g. captured by value) and causing the excessive
memory consumption [2], in case `req` is large like the case of code
update.  This can be verified by doing code update and monitor memory
use between those lines.

```
    getReqAsyncResp->res.setCompleteRequestHandler(std::bind_front(
        afterIfMatchRequest, std::ref(app), asyncResp, req, ifMatch));
```

To address those issues, this commit is suggesting to use `shared_ptr`
to the body content of`Request` so that the copy of `Request` becomes
less overhead (as it is only incrementing ref-count to shared_ptr),
and its content can be valid until all related code references are done.

So, by using `shared_ptr` to the body of `Request', it can captured
by-value for lambda with less overhead (than deep copying).

```
   std::shared_ptr<boost::beast::http::request<bmcweb::HttpBody>> reqPtr;

```

```
     crow::connections::systemBus->async_method_call(
-       [req{std::move(req)}, asyncResp, &rule,
+       [req, asyncResp, &rule, callback(std::forward<CallbackFn>(callback))](

```

- Additionally, there is also an improper move `req` body in
constructor [3].

```
-        auto localReq = std::make_shared<crow::Request>(thisReq.req, ec);
+        auto localReq = std::make_shared<crow::Request>(thisReq.reqPtr);
```

Tested:

- Redfish Validator passes

- Debug logging statement at `completeRequest()` shows the correct `req` value,
  esp it is done on PATCH.
  For example,
```
curl -k -d '{"PowerRestorePolicy":"LastState"}'  -X PATCH https://${bmc}:18080/redfish/v1/Systems/system
```

[1] https://github.com/openbmc/bmcweb/blob/1847f2a0c517df6a7439b9bd491afbf6b2302a92/include/dbus_privileges.hpp#L152C1-L154C56
[2] https://github.com/openbmc/bmcweb/blob/1847f2a0c517df6a7439b9bd491afbf6b2302a92/redfish-core/include/query.hpp#L104C1-L106C1
[3] https://github.com/ibm-openbmc/bmcweb/blob/ef23e48288b35d30baaf57b61299597ca26b1930/redfish-core/include/redfish_aggregator.hpp#L446C1-L448C74

Change-Id: I40a5cc224c1d12e6ee65c99963b9bc6b46379226
Signed-off-by: Myung Bae <myungbae@us.ibm.com>