Rename req var name for clarification
Currently a variable `Body` in `crow::Request` is named as `req` like
```
struct Request
{
    Body req;
```

Also in many places, `crow::Request` variable is referred as `req`. As a
result, this causes the confusion between the references via `req` to
find whether it is `Request` or its body `Request::Body`.

So, Renaming `Request::Body req;` as `Request::Body reqBody;` would help
the readability of the code.

Tested:
- Compiles and runs good
- run_ci and Redfish Service Validator passes

Change-Id: I31ca5d9443adfec26d2caf90867ef3017cdb0a2b
Signed-off-by: Myung Bae <myungbae@us.ibm.com>