Refactor EventLogEntry Get/Collection handling
EventLog Entries are obtained for GET or GET-Collection. However, they
are handled via the different dbus calls
- getAllProperties for GET
- getManagedObjects for Get-Collection

Currently, the content of a log entry is assumed to be under one
interface - `xyz.openbmc_project.Logging.Entry`.

This may cause some difficulties in extending more log contents which
may be under the different interfaces in OEM cases.

This commit is to extract the log contents using the same
getAllProperties() for both GET & GET collection, so that all log
contents can be obtained for the given logEntryPath via
getAllProperties().

Tested:
- Get /redfish/v1/Systems/system/LogServices/EventLog/Entries
- Compare the outputs of before and after change.

Change-Id: Ib68475e9a8ee63af4106bddb84694fad40e334bb
Signed-off-by: Myung Bae <myungbae@us.ibm.com>