Check the dup code update after header read
Currently, if multiple code updates are issued around the same time,
this may cause bmcweb to consume the excessive memory to hold those full
images in the memory, or it may also cause the system out-of-memory
panic.

For example,

- Do the code updates in parallel.
```
   curl  -i  -k -H "X-Auth-Token: $token" -H "Content-Type: application/octet-stream"  -X POST -T ${image}  https://${bmc}/redfish/v1/UpdateService/update &
   curl  -i  -k -H "X-Auth-Token: $token" -H "Content-Type: application/octet-stream"  -X POST -T ${image}  https://${bmc}/redfish/v1/UpdateService/update &
```

- While running the code updates, on BMC system, monitor the memory using `top`.

```
  PID  PPID USER     STAT   VSZ %VSZ CPU %CPU COMMAND
 1167     1 root     R     220m  27%   0   0% /usr/libexec/bmcwebd
 1102     1 root     S    71532   9%   0   0% mboxd --flash 64M --window-size 1M --window-num 63
```

The memory consumptions may increase twice or more depending on the
number of parallel code update attempts.

This is because the current bmcweb reads the entire header & body first.
As soon as the http body is being read, the memory with the body length
is allocated. After that, it goes to the code update logic which may
check whether there is another code update running.
However, at this time, more or too much memory is already allocated.

This commit is to check whether there is already code update running,
right after the http header is read but before body read is started.

Tested:

- Do code updates in parallel.
- The second code update will be rejected as
  `serviceTemporarilyUnavailable`

```
 curl  -i  -k -H "X-Auth-Token: $token" -H "Content-Type: application/octet-stream"  -X POST -T ${image}  https://${bmc}/redfish/v1/UpdateService/Actions/Oem/OemUpdateService.ConcurrentUpdate | tee x1.out
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
  0  202M    0     0    0 1792k      0   507k  0:06:48  0:00:03  0:06:45  507k
HTTP/1.1 503 Service Unavailable
Retry-After: 30
Content-Length: 0
```
- Also monitor the memory consumption on BMC, and its consumption
  remains the same.
- Redfish Service Validator passes

Change-Id: I7a2fcae19c284e7f2771601b3fc6395b270b5c4e
Signed-off-by: Myung Bae <myungbae@us.ibm.com>